requirements.txt
discord.py>=1.2.3
mcstatus>=2.2.1
python-dotenv>=0.10.3